Finishing work

+++ app/Http/Controllers/AdminController.php
@@ -194,25 +194,30 @@ class AdminController extends Controller
             ->exists();
 
 
-        if(!$exists){
-            $newEvent = new Event;
-            $newEvent->place_id = $event->place_id;
-            $newEvent->title = $event->title;
-            $newEvent->about = $event->about;
-            $newEvent->time_from = $event->time_from;
-            $newEvent->time_until = $event->time_until;
-            $newEvent->person_id = $event->person_id;
+        if($start < Carbon\Carbon::now()){
+            if(!$exists){
+                $newEvent = new Event;
+                $newEvent->place_id = $event->place_id;
+                $newEvent->title = $event->title;
+                $newEvent->about = $event->about;
+                $newEvent->time_from = $event->time_from;
+                $newEvent->time_until = $event->time_until;
+                $newEvent->person_id = $event->person_id;
 
-            if($newEvent->save() && $event->delete()){
-                User::find($newEvent->person_id)->notify(new EventAccept($newEvent));
-                return redirect('/admin/events_to_confirm')->with('success', 'Event has been accepted!');
+                if($newEvent->save() && $event->delete()){
+                    User::find($newEvent->person_id)->notify(new EventAccept($newEvent));
+                    return redirect('/admin/events_to_confirm')->with('success', 'Event has been accepted!');
+                }
+                else{
+                    return redirect('/admin/events_to_confirm')->with('error', 'Event has not been accepted...');
+                }
             }
             else{
-                return redirect('/admin/events_to_confirm')->with('error', 'Event has not been accepted...');
+                return redirect('/admin/events_to_confirm')->with('error', 'Event overlaping with other event');
             }
         }
         else{
-            return redirect('/admin/events_to_confirm')->with('error', 'Event overlaping with other event');
+            return redirect('/admin/events_to_confirm')->with('error', 'Cannot add event to past');
         }
